Fixed oracle uint64 to int

+++ app/cmd/root.go
@@ -37,7 +37,7 @@ func addCommands(cmd *cobra.Command) {
 		config.MetricsPort, "metric server port")
 	cmd.PersistentFlags().StringVar(&config.OracleServer, "oracle_server",
 		config.OracleServer, "The oracle server url. (temporary) ")
-	cmd.PersistentFlags().Uint64Var(&config.OracleServerWorldId, "oracle_server_worldid",
+	cmd.PersistentFlags().IntVar(&config.OracleServerWorldId, "oracle_server_worldid",
 		config.OracleServerWorldId, "The worldid to use with the oracle server (temporary) ")
 	cmd.PersistentFlags().StringVar(&config.GenesisTime, "genesis-time",
 		config.GenesisTime, "Time of the genesis layer in 2019-13-02T17:02:00+00:00 format")
